+++ CHANGELOG.md
@@ -48,6 +48,7 @@ You can find its changes [documented below](#060---2020-06-01).
 - `request_timer` can now be called from `LayoutCtx` ([#1278] by [@Majora320])
 - TextBox supports vertical movement ([#1280] by [@cmyr])
 - Widgets can specify a baseline, flex rows can align baselines ([#1295] by [@cmyr])
+- `TextBox::with_text_color` and `TextBox::set_text_color` ([#1320] by [@cmyr])
 
 ### Changed
 
@@ -501,6 +502,7 @@ Last release without a changelog :(
 [#1298]: https://github.com/linebender/druid/pull/1298
 [#1299]: https://github.com/linebender/druid/pull/1299
 [#1311]: https://github.com/linebender/druid/pull/1311
+[#1320]: https://github.com/linebender/druid/pull/1320
 
 [Unreleased]: https://github.com/linebender/druid/compare/v0.6.0...master
 [0.6.0]: https://github.com/linebender/druid/compare/v0.5.0...v0.6.0